chore: Move environment functions to environment_utils
