+++ source/backend/executor/executor_utils.c
@@ -31,41 +31,6 @@ char	*get_base_filename(char	*full_name)
 	return (ret);
 }
 
-char	*get_value_with_key(char *str, char *key)
-{
-	int		key_len;
-	char	*value;
-
-	if (!str || !key)
-		return (NULL);
-	key_len = ft_strlen(key);
-	if (ft_strncmp(str, key, key_len) == 0 && str[key_len] == '=')
-	{
-		value = str + key_len + 1;
-		return (value);
-	}
-	return (NULL);
-}
-
-char	*get_value_from_env(char **envp, char *key)
-{
-	int		i;
-	char	*value;
-
-	if (!envp || !key)
-		return (NULL);
-	value = NULL;
-	i = 0;
-	while (envp[i])
-	{
-		value = get_value_with_key(envp[i], key);
-		if (value)
-			break ;
-		i++;
-	}
-	return (value);
-}
-
 bool	set_all_path(char ***all_path, char **envp)
 {
 	char	*path_value;
